+++ influxdb_client_3/write_client/client/influxdb_client.py
@@ -293,6 +293,10 @@ class InfluxDBClient(_BaseClient):
 
     def __del__(self):
         """Shutdown the client."""
-        if self.api_client:
-            self.api_client.__del__()
-            self.api_client = None
+        try:
+            if self.api_client:
+                self.api_client.__del__()
+                self.api_client = None
+        except (TypeError, AttributeError):
+            # During interpreter shutdown, cleanup may fail
+            pass